factor Stat out into its own module
